refactor: minor changes + added ChartSettings.ts
